String  oke
L'axe X affiche maintenant les jour de la semaine

+++ App/NBAStatsCalculator/Form1.cs
@@ -3,7 +3,7 @@ namespace NBAStatsCalculator
     public partial class Form1 : Form
     {
         public  double[] dayOfWeek = {1, 2, 3, 4, 5, 6, 7 };
-        public int[] nbPoints = { 10, 20, 30, 40, 50, 60, 70 };
+        public double[] nbPoints = { 10, 20, 30, 40, 50, 60, 70 };
         public DateTime dateTest = new DateTime(2024, 9, 6);
         public Form1()
         {
